requirements.txt
nltk
sentencepiece
requests
tqdm
patool
